+++ Sources/LiveLocalizationCore/String+LiveLocalization.swift
@@ -2,18 +2,11 @@ import Foundation
 
 public extension String {
     func localize() async -> String {
-        await LiveLocalizer.shared.localize(self)
-    }
-
-    func localizeSync() -> String {
-        LiveLocalizer.shared.localizeSync(self)
+        let localizer = await LiveLocalization.localizer
+        return await localizer.localize(self)
     }
 
     func localize(using localizer: LiveLocalizer) async -> String {
         await localizer.localize(self)
-    }
-
-    func localizeSync(using localizer: LiveLocalizer) -> String {
-        localizer.localizeSync(self)
     }
 }